changed Telephone type to bigint in tables

+++ db/migrate/20170724080011_hosts.rb
@@ -2,7 +2,7 @@ class Hosts < ActiveRecord::Migration[5.1]
   def change
   	create_table(:hosts) do |t|
   		t.column(:name, :string)
-  		t.column(:Telephone, :integer)
+  		t.column(:Telephone, :bigint)
   		t.column(:Email, :varchar)
 
   		t.timestamps()
